1.0.0 a3

+++ core/Initializer.py
@@ -1,33 +1,32 @@
 import getpass
 import importlib
 import json
-import logging
-import shutil
+import socket
 import subprocess
 import time
 from pathlib import Path
 
 import yaml
 
-from core.commons import commons
+import configTemplate
+from core.base.model.ProjectAliceObject import ProjectAliceObject
 
 
-class initDict(dict):
+class initDict(dict, ProjectAliceObject):
 
 	def __init__(self, default: dict):
-		self._logger = logging.getLogger('ProjectAlice')
-		super().__init__(default)
+		super().__init__(default, logDepth=3)
 
 
 	def __getitem__(self, item):
 		try:
 			return super().__getitem__(item) or ''
 		except:
-			self._logger.warning(f'Missing key "{item}" in provided yaml file. Are you using a deprecated yaml file version?')
+			self.logWarning(f'Missing key "{item}" in provided yaml file. Are you using a deprecated yaml file version?')
 			return ''
 
 
-class Initializer:
+class Initializer(ProjectAliceObject):
 
 	NAME = 'ProjectAlice'
 
@@ -44,19 +43,22 @@ network={
 	'''
 
 	def __init__(self):
-		self._logger = logging.getLogger('ProjectAlice')
-		self._logger.info('Starting Project Alice initializer')
+		super().__init__(logDepth=3)
+		self.logInfo('Starting Project Alice initializer')
 
-		self._confsFile = Path(commons.rootDir(), 'config.py')
-		self._confsSample = Path(commons.rootDir(), 'configSample.py')
+		self._rootDir = Path(__file__).resolve().parent.parent
+
+		self._confsFile = Path(self._rootDir, 'config.py')
+		self._confsSample = Path(self._rootDir, 'configTemplate.py')
 		self._initFile = Path('/boot/ProjectAlice.yaml')
-		self._latest = 1.03
+		self._latest = 1.07
 
 
 	def initProjectAlice(self) -> bool:
 		if not self._initFile.exists() and not self._confsFile.exists():
 			self.fatal('Init file not found and there\'s no configuration file, aborting Project Alice start')
 		elif not self._initFile.exists():
+			self.logInfo('No initialization needed')
 			return False
 
 		with self._initFile.open(mode='r') as f:
@@ -69,38 +71,49 @@ network={
 		if float(initConfs['version']) < self._latest:
 			self.fatal('The yaml file you are using is deprecated. Please update it before trying again')
 
-		# Let's connect to wifi!
-		if not initConfs['wifiCountryCode'] or not initConfs['wifiNetworkName'] or not initConfs['wifiWPAPass']:
-			self.fatal('You must specify the wifi parameters')
-
 		wpaSupplicant = Path('/etc/wpa_supplicant/wpa_supplicant.conf')
-		bootWpaSupplicant = Path('/boot/wpa_supplicant.conf')
+		if not wpaSupplicant.exists() and initConfs['useWifi']:
+			self.logInfo('Setting up wifi')
 
-		if not wpaSupplicant.exists():
-			self._logger.info('Setting up wifi')
+			if not initConfs['wifiCountryCode'] or not initConfs['wifiNetworkName'] or not initConfs['wifiWPAPass']:
+				self.fatal('You must specify the wifi parameters')
+
+			bootWpaSupplicant = Path('/boot/wpa_supplicant.conf')
+
 			wpaFile = self._WPA_FILE\
-				.replace('%wifiCountryCode%', initConfs['wifiCountryCode'])\
-				.replace('%wifiNetworkName%', initConfs['wifiNetworkName'])\
-				.replace('%wifiWPAPass%', initConfs['wifiWPAPass'])
+				.replace('%wifiCountryCode%', str(initConfs['wifiCountryCode']))\
+				.replace('%wifiNetworkName%', str(initConfs['wifiNetworkName']))\
+				.replace('%wifiWPAPass%', str(initConfs['wifiWPAPass']))
 
-			file = Path(commons.rootDir(), 'wifi.conf')
+			file = Path(self._rootDir, 'wifi.conf')
 			file.write_text(wpaFile)
 
-			self._logger.info('wpa_supplicant.conf')
 			subprocess.run(['sudo', 'mv', str(file), bootWpaSupplicant])
+			self.logInfo('Successfully initialized wpa_supplicant.conf')
 			time.sleep(1)
 			subprocess.run(['/usr/bin/sudo', '/sbin/shutdown', '-r', 'now'])
 			exit(0)
+
+		connected = False
+		try:
+			socket.create_connection(('www.google.com', 80))
+			connected = True
+		except:
+			pass
+
+		if not connected:
+			self.fatal('Your device needs internet access to continue, to download the updates and create the assistant')
 
 		if not initConfs['snipsConsoleLogin'] or not initConfs['snipsConsolePassword'] or not initConfs['intentsOwner']:
 			self.fatal('You must specify a Snips console login, password and intent owner')
 
 		if not self._confsFile.exists() and not self._confsSample.exists():
-			self.fatal('No config and no config sample found, can\'t continue')
+			self.fatal('No config and no config template found, can\'t continue')
 
 		elif not self._confsFile.exists() and self._confsSample.exists():
 			self.warning('No config file found, creating it from sample file')
-			shutil.copyfile(src=Path(commons.rootDir(), 'configSample.py'), dst=Path(commons.rootDir(), 'config.py'))
+			confs = {configName: configData['defaultValue'] if 'defaultValue' in configData else configData for configName, configData in configTemplate.settings.items()}
+			Path('config.py').write_text(f"settings = {json.dumps(confs, indent=4).replace('false', 'False').replace('true', 'True')}")
 
 		elif self._confsFile.exists() and not initConfs['forceRewrite']:
 			self.warning('Config file already existing and user not wanting to rewrite, aborting')
@@ -108,8 +121,9 @@ network={
 
 		elif self._confsFile.exists() and initConfs['forceRewrite']:
 			self.warning('Config file found and force rewrite specified, let\'s restart all this!')
-			Path(commons.rootDir(), 'config.py').unlink()
-			shutil.copyfile(src=Path(commons.rootDir(), 'configSample.py'), dst=Path(commons.rootDir(), 'config.py'))
+			Path(self._rootDir, 'config.py').unlink()
+			confs = {configName: configData['defaultValue'] if 'defaultValue' in configData else configData for configName, configData in configTemplate.settings.items()}
+			Path('config.py').write_text(f"'settings = {json.dumps(confs, indent=4).replace('false', 'False').replace('true', 'True')}'")
 
 		config = importlib.import_module('config')
 		confs = config.settings.copy()
@@ -122,16 +136,18 @@ network={
 		subprocess.run(['git', 'pull'])
 		subprocess.run(['git', 'stash', 'clear'])
 
-
+		subprocess.run(['./venv/bin/pip3', 'uninstall', '-r', str(Path(self._rootDir, 'pipuninstalls.txt'))])
 		# Do some installation if wanted by the user
 		if initConfs['doGroundInstall']:
+			subprocess.run(['./venv/bin/pip3', 'install', '-r', str(Path(self._rootDir, 'piprequirements.txt'))])
+
 			subprocess.run(['sudo', 'bash', '-c', 'echo "deb https://raspbian.snips.ai/$(lsb_release -cs) stable main" > /etc/apt/sources.list.d/snips.list'])
 			subprocess.run(['sudo', 'apt-key',  'adv', '--keyserver', 'gpg.mozilla.org', '--recv-keys', 'D4F50CDCA10A2849'])
 			subprocess.run(['sudo', 'apt-get', 'update'])
 
-			reqs = [line.rstrip('\n') for line in open(Path(commons.rootDir(), 'sysrequirements.txt'))]
+			reqs = [line.rstrip('\n') for line in open(Path(self._rootDir, 'sysrequirements.txt'))]
 			subprocess.run(['sudo', 'apt-get', 'install', '-y', '--allow-unauthenticated'] + reqs)
-			subprocess.run(['./venv/bin/pip3', 'install', '-r', str(Path(commons.rootDir(), 'piprequirements.txt'))])
+
 			subprocess.run(['sudo', 'systemctl', 'stop', 'snips-*'])
 			subprocess.run(['sudo', 'systemctl', 'disable', 'snips-asr'])
 			subprocess.run(['sudo', 'systemctl', 'disable', 'snips-nlu'])
@@ -145,7 +161,7 @@ network={
 		# Now let's dump some values to their respective places
 		# First those that need some checks and self filling in case unspecified
 		confs['mqttHost'] = str(initConfs['mqttHost']) or 'localhost'
-		confs['mqttPort'] = str(initConfs['mqttPort']) or '1883'
+		confs['mqttPort'] = initConfs['mqttPort'] or 1883
 
 		confs['snipsConsoleLogin'] = initConfs['snipsConsoleLogin']
 		confs['snipsConsolePassword'] = initConfs['snipsConsolePassword']
@@ -172,7 +188,7 @@ network={
 			confs['awsSecretKey'] = initConfs['awsSecretKey']
 
 			if initConfs['googleServiceFile']:
-				googleCreds = Path(commons.rootDir(), 'credentials/googlecredentials.json')
+				googleCreds = Path(self._rootDir, 'credentials/googlecredentials.json')
 				googleCreds.write_text(json.dumps(initConfs['googleServiceFile']))
 
 		# Those that don't need checking
@@ -182,6 +198,7 @@ network={
 		confs['micChannels'] = int(initConfs['micChannels'])
 		confs['useSLC'] = bool(initConfs['useSLC'])
 		confs['webInterfaceActive'] = bool(initConfs['webInterfaceActive'])
+		confs['webInterfaceDevMode'] = bool(initConfs['webInterfaceDevMode'])
 		confs['newDeviceBroadcastPort'] = int(initConfs['newDeviceBroadcastPort'])
 		confs['activeLanguage'] = initConfs['activeLanguage']
 		confs['activeCountryCode'] = initConfs['activeCountryCode']
@@ -198,6 +215,7 @@ network={
 		confs['githubUsername'] = initConfs['githubUsername']
 		confs['githubToken'] = initConfs['githubToken']
 		confs['ttsLanguage'] = initConfs['ttsLanguage']
+		confs['updateChannel'] = initConfs['updateChannel']
 
 		if initConfs['snipsProjectId'] and confs['activeLanguage'] in confs['supportedLanguages']:
 			confs['supportedLanguages'][confs['activeLanguage']]['snipsProjectId'] = initConfs['snipsProjectId']
@@ -205,7 +223,7 @@ network={
 		if initConfs['deviceName'] != 'default':
 			subprocess.run(['sudo', 'sed', '-i', '-e', f's/\# bind = "default@mqtt"/bind = "{initConfs["deviceName"]}@mqtt"/', Path('/etc/snips.toml')])
 			subprocess.run(['sudo', 'sed', '-i', '-e', f's/bind = ".*@mqtt"/bind = "{initConfs["deviceName"]}@mqtt"/', Path('/etc/snips.toml')])
-			subprocess.run(['sudo', 'sed', '-i', '-e', f's/DEFAULT_SITE_ID = \'default\'/DEFAULT_SITE_ID = \'{initConfs["deviceName"]}\'/', Path(commons.rootDir(), 'core/commons/constants.py')])
+			subprocess.run(['sudo', 'sed', '-i', '-e', f's/DEFAULT_SITE_ID = \'default\'/DEFAULT_SITE_ID = \'{initConfs["deviceName"]}\'/', Path(self._rootDir, 'core/commons/constants.py')])
 
 		serviceFilePath = Path('/etc/systemd/system/ProjectAlice.service')
 		if not serviceFilePath.exists():
@@ -215,7 +233,7 @@ network={
 		subprocess.run(['sudo', 'sed', '-i', '-e', f's/\#EXECSTART/ExecStart=\/home\/{getpass.getuser()}\/ProjectAlice\/venv\/bin\/python3 main.py/', str(serviceFilePath)])
 		subprocess.run(['sudo', 'sed', '-i', '-e', f's/\#USER/User={getpass.getuser()}/', str(serviceFilePath)])
 
-		self._logger.info('Installing audio hardware')
+		self.logInfo('Installing audio hardware')
 		audioHardware = ''
 		for hardware in initConfs['audioHardware']:
 			if initConfs['audioHardware'][hardware]:
@@ -240,35 +258,39 @@ network={
 			subprocess.run(['sudo', 'sed', '-i', '-e', f's/%USER%/{getpass.getuser()}/', str(slcServiceFilePath)])
 
 		if audioHardware in {'respeaker2', 'respeaker4', 'respeaker6MicArray'}:
-			subprocess.run(['sudo', Path(commons.rootDir(), 'system/scripts/audioHardware/respeakers.sh')])
+			subprocess.run(['sudo', Path(self._rootDir, 'system/scripts/audioHardware/respeakers.sh')])
 			if initConfs['useSLC']:
 				subprocess.run(['sudo', 'sed', '-i', '-e', f's/%HARDWARE%/{audioHardware}/', str(slcServiceFilePath)])
 
 			if audioHardware == 'respeaker6MicArray':
-				subprocess.run(['sudo', 'cp', Path(commons.rootDir(), 'system', 'asounds', 'respeaker6micarray.conf'), Path('/etc/asound.conf')])
+				subprocess.run(['sudo', 'cp', Path(self._rootDir, 'system', 'asounds', 'respeaker6micarray.conf'), Path('/etc/asound.conf')])
 
 		elif audioHardware == 'respeaker7':
-			subprocess.run(['sudo', Path(commons.rootDir(), 'system/scripts/audioHardware/respeaker7.sh')])
+			subprocess.run(['sudo', Path(self._rootDir, 'system/scripts/audioHardware/respeaker7.sh')])
 			if initConfs['useSLC']:
 				subprocess.run(['sudo', 'sed', '-i', '-e', 's/%HARDWARE%/respeaker7MicArray/', str(slcServiceFilePath)])
 
 		elif audioHardware == 'respeakerCoreV2':
-			subprocess.run(['sudo', Path(commons.rootDir(), 'system/scripts/audioHardware/respeakerCoreV2.sh')])
+			subprocess.run(['sudo', Path(self._rootDir, 'system/scripts/audioHardware/respeakerCoreV2.sh')])
 			if initConfs['useSLC']:
 				subprocess.run(['sudo', 'sed', '-i', '-e', f's/%HARDWARE%/{audioHardware}/', str(slcServiceFilePath)])
 
 		elif audioHardware in {'matrixCreator', 'matrixVoice'}:
-			subprocess.run(['sudo', Path(commons.rootDir(), 'system/scripts/audioHardware/matrix.sh')])
+			subprocess.run(['sudo', Path(self._rootDir, 'system/scripts/audioHardware/matrix.sh')])
+			subprocess.run(['sudo', 'cp', Path(self._rootDir, 'system', 'asounds', 'matrix.conf'), Path('/etc/asound.conf')])
+
+			subprocess.run(['sudo', 'sed', '-i', '-e', f's/\# mike = "Built-in Microphone"/mike = "MATRIXIO-SOUND: - (hw:2,0)"/', Path('/etc/snips.toml')])
+
 			if initConfs['useSLC']:
 				subprocess.run(['sudo', 'sed', '-i', '-e', f's/%HARDWARE%/{audioHardware.lower()}/', str(slcServiceFilePath)])
 
 		elif audioHardware == 'googleAIY':
-			subprocess.run(['sudo', Path(commons.rootDir(), 'system/scripts/audioHardware/aiy.sh')])
+			subprocess.run(['sudo', Path(self._rootDir, 'system/scripts/audioHardware/aiy.sh')])
 			if initConfs['useSLC']:
 				subprocess.run(['sudo', 'sed', '-i', '-e', 's/%HARDWARE%/googleAIY/', str(slcServiceFilePath)])
 
 		elif audioHardware == 'usbMic':
-			subprocess.run(['sudo', 'cp', Path(commons.rootDir(), 'system', 'asounds', 'usbmic.conf'), Path('/etc/asound.conf')])
+			subprocess.run(['sudo', 'cp', Path(self._rootDir, 'system', 'asounds', 'usbmic.conf'), Path('/etc/asound.conf')])
 
 		subprocess.run(['sudo', 'systemctl', 'daemon-reload'])
 
@@ -284,9 +306,9 @@ network={
 		else:
 			importlib.reload(config)
 
-		subprocess.run(['sudo', 'rm', '-rf', Path(commons.rootDir(), 'assistant')])
-		subprocess.run(['sudo', 'rm', '-rf', Path(commons.rootDir(), 'trained', 'assistants', f"assistant_{confs['activeLanguage']}")])
-		subprocess.run(['sudo', 'rm', '-rf', Path(commons.rootDir(), 'var', 'assistants', confs['activeLanguage'])])
+		subprocess.run(['sudo', 'rm', '-rf', Path(self._rootDir, 'assistant')])
+		subprocess.run(['sudo', 'rm', '-rf', Path(self._rootDir, 'trained', 'assistants', f"assistant_{confs['activeLanguage']}")])
+		subprocess.run(['sudo', 'rm', '-rf', Path(self._rootDir, 'var', 'assistants', confs['activeLanguage'])])
 		subprocess.run(['sudo', 'mv', str(Path('/boot/ProjectAlice.yaml')), str(Path('/boot/ProjectAlice.yaml.bak'))])
 
 		self.warning('Initializer done with configuring')
@@ -296,9 +318,9 @@ network={
 
 
 	def fatal(self, text: str):
-		self._logger.fatal(text)
+		self.logFatal(text)
 		exit()
 
 
 	def warning(self, text: str):
-		self._logger.warning(text)
+		self.logWarning(text)
